more rafctoring+unit tests

+++ src/controller/ReturnController.ts
@@ -40,38 +40,46 @@ export class ReturnController {
 
     async handleRedirect(req: any, res: any): Promise<any> {
 
-    	if (req.query.error) {
-    		loggingHelper.error("Received error response from /authorize", {
-    			"error": req.query.error,
-    			"error_description": req.query.error_description,
-    		});
-    		loggingHelper.info("Redirecting to accounts dashboard");
-    		res.redirect(EnvironmentVariables.getAccountsDashboardUrl());
-    	}
-    	if (req.query.state && req.query.code) {
-    		try {
-    			loggingHelper.info("Received success response", { "code": req.query.code, "state": req.query.state });
-    			//await ReturnController.getInstance().handleRedirect(req.query.state as string)
-    			await this.iprService.deleteSession(req.query.state);
-    		} catch {
-    			loggingHelper.error("Got error deleting record from DB");
-    			loggingHelper.info("Redirecting to accounts dashboard");
-    			res.redirect(EnvironmentVariables.getAccountsDashboardUrl());
-    		}
+		if(req.query){
+			if (req.query.error) {
+				loggingHelper.error("Received error response from /authorize", {
+					"error": req.query.error,
+					"error_description": req.query.error_description,
+				});
+				this.redirectToDashboard(res,"Received error response from /authorize");
+			}
+			if (req.query.state && req.query.code) {
+				try {
+					loggingHelper.info("Received success response", { "code": req.query.code, "state": req.query.state });
+					//await ReturnController.getInstance().handleRedirect(req.query.state as string)
+					await this.iprService.deleteSession(req.query.state);
+				} catch {
+					this.redirectToDashboard(res,"Got error deleting record from DB");
+				}
 
-    		try {
-    			const resp = await axios.get(`${EnvironmentVariables.getApiBaseUrl()}/session?code=${req.query.code}`);
-    			console.log("Redirecting to ", resp.data?.redirect_uri);
-    			res.redirect(resp.data?.redirect_uri);
-    		} catch (e) {
-    			loggingHelper.error("Received error calling /session and redirecting to RP", { "error": e });
-    			loggingHelper.info("Redirecting to accounts dashboard");
-    			res.redirect(EnvironmentVariables.getAccountsDashboardUrl());
-    		}
-    	} else {
-    		loggingHelper.error("Missing mandatory fields");
-    		loggingHelper.info("Redirecting to accounts dashboard");
-    		res.redirect(EnvironmentVariables.getAccountsDashboardUrl());
-    	}
+				try {
+					const resp = await axios.get(`${EnvironmentVariables.getApiBaseUrl()}/session?code=${req.query.code}`);
+					loggingHelper.info("Redirecting to ", {"redirectUri":resp.data?.redirect_uri});
+					res.redirect(resp.data?.redirect_uri);
+				} catch (e) {
+					this.redirectToDashboard(res,"Received error calling /session and redirecting to RP", e );
+				}
+			} else {
+				this.redirectToDashboard(res, "Missing mandatory fields");
+			}
+		} else{
+			this.redirectToDashboard(res, "Missing query parameters in request");
+		}
+
     }
+
+	redirectToDashboard(res: any, reason: string, error?: any){
+		if(error){
+			loggingHelper.error("Redirecting to accounts dashboard", {"reason": reason, "error":error});
+		}else{
+			loggingHelper.error("Redirecting to accounts dashboard", {"reason": reason});
+		}
+
+		res.redirect(EnvironmentVariables.getAccountsDashboardUrl());
+	}
 }
